Major modification of AST structure, to remove "sharing" of nodes Method, Predicate, Function in nodes such as MethodCall, PredicateAccess, FuncApp etc.
The reason is that these "shared" nodes are not treated correctly when the AST is transformed; if the whole program is transformed, then old copies of these nodes will remain in MethodCall etc. nodes, because the transformer code skips recursing over these to avoid infinite recursion.

In general, the fix is to replace the occurrences of these nodes with string identifiers. For example, a MethodCall node now does *not* store a Method node as a child, but only the method name as a String. This must be looked-up in the current Program, to find the appropriate Method node.

To minimise the changes needed, and also to allow some consistency checks to still be made, when *creating* such nodes, the signature hasn't changed. Thus, a MethodCall can still be created passing a Method node as the first parameter. However, this node will not be captured during creation (only its name etc. will be saved)

To facilitate the extra code for looking up method names etc., Program now contains methods findFunction, findMethod etc., which take a string as name and return the corresponding node (or an error if not found)


Other changes:

FuncApp and MethodCall no longer have "pres" and "posts" as lazy vals (because they should be calculated from the program definitions); these must be computed from the appropriate looked-up Function/Method "pres"/"posts" fields. e.g. instead of fa.pres one should write
func = program.findFunction(fa.name);
func.pres map (e => Expressions.instantiateVariables(e, func.formalArgs, fa.args))

The code in Transformer.scala no longer skips Field nodes as childs of e.g. a FieldAccess node (this could cause transformations on types to not be applied to these nodes). As yet, the overall change made for Method/Function nodes has not been applied to Field nodes; this is less important, since there is no need for the Transformer to skip them as children (no potential infinite recursion, unlike for methods with bodies)

loc in the LocationAccess trait is now parameterised by a Program (so that predicate locations can be calculated based on the looked-up Predicate node)
predicateBody in PredicateAccess is parameterised similarly

func in FuncLikeApp also returns a closure of type (Program => FuncLike), but this code is likely to be modified again very soon (to apply similar changes for domain functions)

unapply in FuncLikeApp now gives a pair of a string (name) and args sequence, not a FuncLike and args.

callee in trait Call is now of type String and not Callable (again, this entails a lookup in client code)

proofObligations in ast/utility/Expressions.scala now returns a closure requiring a Program (to lookup function preconditions in)



Orthogonal change: fixed a bug in the "desugar" method of ExplicitSeq (but I don't think anyone else was using this code yet, anyway)